requirements.txt
Django==4.2.7
django-filter==23.4
djangorestframework==3.14.0
Markdown==3.5.1
celery==5.3.6
django-celery-results==2.5.1
redis==5.0.1
flower==2.0.1
django-redis==5.4.0
django-extensions==3.2.3
